Move GATModel to covid_moonshot_ml.nn.models.GAT.
