+++ covid_moonshot_ml/nn/models/__init__.py
@@ -1,4 +1,5 @@
 from .E3NNBind import E3NNBind
+from .GAT import GAT
 from .SchNetBind import SchNetBind
 
-__all__ = ["E3NNBind", "SchNetBind"]
+__all__ = ["E3NNBind", "GAT", "SchNetBind"]
